+++ test/widgets/panel_edge_cases_test.dart
@@ -1,6 +1,7 @@
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:panel_layout/panel_layout.dart';
+import 'package:panel_layout/src/widgets/panel_resize_handle.dart';
 
 void main() {
   group('Edge Cases', () {